Organized the codes into folders :D
